+++ src/yb/consensus/opid_util.h
@@ -107,7 +107,7 @@ struct OpIdCompareFunctor {
 };
 
 // OpId comparison functor that returns true iff left > right. Suitable for use
-// td::sort and std::map to sort keys in increasing order.]
+// with std::sort and std::map to sort keys in increasing order.]
 struct OpIdBiggerThanFunctor {
   bool operator() (const OpId& left, const OpId& right) const;
 };
@@ -124,7 +124,7 @@ std::string OpIdToString(const OpId& id);
 
 std::string OpsRangeString(const ConsensusRequestPB& req);
 
-OpId MakeOpId(int term, int index);
+OpId MakeOpId(int64_t term, int64_t index);
 
 }  // namespace consensus
